Fix initialize active_in_last_period to false
This way activity log is not sent immediately but after the first real action occurs. When I have more tabs open and then restart browser it loads them again. With this fix the page visit is not logged with time spent > 0 before I actually do some actions on the page.